doc_scraper/requirements.txt
# Documentation Scraper Requirements

# Core scraping libraries
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0
lxml>=4.9.3

# HTML to text conversion
html2text>=2020.1.16
markdown>=3.5

# Progress bars and logging
tqdm>=4.66.1
colorlog>=6.8.0

# Configuration
pyyaml>=6.0.1

# Search and indexing
whoosh>=2.7.4
rich>=13.7.0

# Browser driver management (for Selenium)
webdriver-manager>=4.0.1

# Rate limiting
ratelimit>=2.2.1

# Authentication helpers
requests-oauthlib>=1.3.1

# Optional: For parallel scraping
aiohttp>=3.9.0
asyncio>=3.4.3

# Optional: For better text extraction
trafilatura>=1.6.2

# Optional: For PDF documentation
PyPDF2>=3.0.1
